implement animations
currently broken for the CTA section
