+++ package.json
@@ -16,6 +16,7 @@
     "react-resizable": "^3.0.4",
     "react-rnd": "^10.3.3",
     "react-scripts": "4.0.3",
+    "rellax": "^1.12.1",
     "typescript": "^4.3.4",
     "web-vitals": "^1.1.2"
   },
@@ -48,6 +49,7 @@
   },
   "devDependencies": {
     "@types/react-resizable": "^1.7.2",
+    "@types/rellax": "^1.7.4",
     "@typescript-eslint/eslint-plugin": "^4.28.0",
     "@typescript-eslint/parser": "^4.28.0",
     "autoprefixer": "^10.2.6",
