+++ src/components/footer/callToAction.tsx
@@ -1,22 +1,42 @@
-import React from "react";
-import Sun from "../../assets/callToAction/sun.svg";
-import { ReactComponent as CityView } from "../../assets/callToAction/cityView.svg";
+import React, {useEffect} from "react";
 import "../common.scss";
+import Rellax from "rellax";
 
 export const CallToAction: React.FC = () => {
   const goToRepo = (): void => {
     window.open("https://github.com/eugenefrg/sunset-hills", "_blank");
   };
 
+  useEffect(() => {
+    new Rellax(".layer1");
+    new Rellax(".layer2");
+    new Rellax(".layer3");
+  },[]);
   return (
     <div
-      className="bg-orange h-3/5 bg-top bg-no-repeat"
-      style={{ backgroundImage: `url(${Sun})`, backgroundSize: "100% " }}
+      className="h-screen relative overflow-hidden svgContainer"
     >
-      <div className="h-3/5 absolute">
-        <CityView />
-      </div>
-      <div className="absolute h-3/5 w-full">
+      <svg className="absolute" viewBox="0 0 1920 1080" fill="none" xmlns="http://www.w3.org/2000/svg">
+        <g clipPath="url(#clip0)">
+          <path d="M0.00012207 1299L1920 1299L1920 1044.41L9.98133e-05 1044.41L0.00012207 1299Z" fill="#E6645B"/>
+          <path d="M0.00012207 754.906L1920 754.905L1920 -219L3.69288e-05 -219L0.00012207 754.906Z" fill="#FDD05F"/>
+          <path d="M0.00012207 1044.41L1920 1044.41L1920 754.905L9.6761e-05 754.906L0.00012207 1044.41Z" fill="#FFA163"/>
+          <path d="M427.821 1155.33C462.381 1188.51 502.588 1214.53 547.426 1232.75C593.957 1251.62 643.313 1261.27 694.136 1261.27C744.96 1261.27 794.202 1251.73 840.847 1232.75C885.684 1214.53 926.004 1188.51 960.451 1155.33C995.011 1122.15 1022.12 1083.55 1041.09 1040.51C1060.74 995.834 1070.79 948.45 1070.79 899.658C1070.79 850.865 1060.86 803.59 1041.09 758.809C1022.12 715.763 995.011 677.054 960.451 643.983C925.891 610.804 885.684 584.781 840.847 566.565C794.315 547.698 744.96 538.048 694.136 538.048C643.313 538.048 594.07 547.59 547.425 566.565C502.588 584.781 462.268 610.804 427.821 643.983C393.261 677.162 366.155 715.763 347.181 758.809C327.529 803.481 317.477 850.865 317.477 899.658C317.477 948.45 327.416 995.725 347.181 1040.51C366.155 1083.55 393.374 1122.26 427.821 1155.33ZM694.249 754.905C777.487 754.905 845.026 819.746 845.026 899.658C845.026 979.569 777.487 1044.41 694.249 1044.41C611.011 1044.41 543.473 979.569 543.473 899.658C543.473 819.746 610.898 754.905 694.249 754.905Z" fill="#FDD05F"/>
+          <path d="M694.249 754.905C610.977 754.905 543.473 819.713 543.473 899.658C543.473 979.602 610.977 1044.41 694.249 1044.41C777.52 1044.41 845.025 979.602 845.025 899.658C845.025 819.713 777.52 754.905 694.249 754.905Z" fill="white"/>
+          <path d="M1041.21 758.809C1060.86 803.481 1070.91 850.865 1070.91 899.658C1070.91 948.45 1060.97 995.725 1041.21 1040.51C1022.23 1083.55 995.125 1122.26 960.565 1155.33C926.005 1188.51 885.798 1214.53 840.96 1232.75C794.429 1251.62 745.073 1261.27 694.25 1261.27C643.426 1261.27 594.184 1251.73 547.539 1232.75C502.702 1214.53 462.382 1188.51 427.935 1155.33C393.375 1122.15 366.269 1083.55 347.295 1040.51C327.643 995.834 317.591 948.45 317.591 899.658C317.591 850.865 327.53 803.59 347.295 758.809C347.859 757.508 348.537 756.207 349.102 754.905L110.796 754.905C98.1462 801.963 91.7087 850.431 91.7087 899.658C91.7087 977.509 107.746 1053.41 139.257 1124.97C167.831 1189.7 207.812 1248.26 258.297 1299L1130.2 1299C1180.69 1248.26 1220.78 1189.7 1249.24 1124.97C1280.75 1053.41 1296.79 977.618 1296.79 899.658C1296.79 850.431 1290.35 801.963 1277.7 754.905L1039.4 754.905C1039.96 756.207 1040.64 757.508 1041.21 758.809Z" fill="#FFA163"/>
+          <path className="layer3" data-rellax-speed="-3" d="M1801.41 795.374L1724.61 795.374L1724.61 900.442L1708.8 900.442L1708.8 795.374L1629.74 795.374L1629.74 566.59L1584.56 566.59L1584.56 761.762L1561.98 761.762L1561.98 486.353L1541.65 486.353L1541.65 648.996L1482.92 648.996L1482.92 568.759L1435.48 568.759L1435.48 872.359L1410.64 896.213L1410.64 559L1356.42 559L1356.42 855.01L1270.59 855.01L1270.59 470.089L1211.86 470.089L1211.86 615.708L1037.93 615.708L1037.93 510.207L789.459 510.207L789.459 615.708L651.671 615.708L651.671 492.859L459.671 492.859L459.671 659.839L389.647 659.839L389.647 415.874L358.024 415.874L358.024 714.053L303.812 766.099L303.812 490.69L265.412 490.69L265.412 857.179L150.212 857.179L150.212 640.322L96.0001 640.322L96.0001 750.919L0.000101369 750.919L1.82373e-05 -200L1920 -200L1920 795.374L1801.41 795.374Z" fill="#E7655B"/>
+          <path className="layer2" data-rellax-speed="-5" d="M1920 730.615L1776.45 818.442L1776.45 354.91L1749.35 354.91L1749.35 415.087L1684.97 415.087L1684.97 379.306L1613.82 379.306L1518.95 379.306L1518.95 325.634L1459.65 325.634L1459.65 633.029L1386.8 633.029L1386.8 346.777L1359.7 346.777L1319.04 592.368L1268.22 286.6L1229.25 168.521L1176.73 168.521L1176.73 607.006L1137.77 607.006L1137.77 631.402L1049.68 631.402L1049.68 608.632L998.852 608.632L998.852 388.089L898.899 388.089L898.899 504.541L878.569 504.541L878.569 388.089L831.134 388.089L780.311 319.128L705.769 319.128L705.769 551.707L544.828 551.708L544.828 263.83L334.758 263.83L334.758 444.363L224.64 444.363L224.64 262.203L165.346 262.203L165.346 509.42L51.8401 509.42L51.8401 566.345L9.8516e-05 636.282L1.82373e-05 -282L1920 -282L1920 730.615Z" fill="#9B3467"/>
+          <path className="layer1" data-rellax-speed="-7" d="M1920 419.348L1828.97 474.864L1712.08 419.565L1712.08 256.922L1632.45 256.922L1632.45 198.371L1568.08 198.371L1568.08 156.084L1512.17 156.084L1512.17 305.715L1385.11 305.715L1339.37 284.572L1339.37 71.5095L1310.57 71.5095L1310.57 43.8602L1286.85 43.8602L1286.85 248.79L1139.46 248.79L1073.39 213.009L1073.39 53.6188L960 53.6188L960 292.27L738.635 292.27L738.635 193.6L728.471 193.6L728.471 340.521L639.247 505.874L552.282 340.521L552.282 274.922L472.094 274.922L472.094 352.99L456.282 352.99L456.282 371.423L303.812 371.423L303.812 354.074L288 354.074L288 287.933L100.518 287.933L100.518 243.477L44.0471 243.477L44.0471 258.657L9.28285e-05 258.657L6.05995e-05 -110L1920 -110L1920 419.348Z" fill="#39193C"/>
+        </g>
+        <defs>
+          <clipPath id="clip0">
+            <rect width="1920" height="1080" fill="white"/>
+          </clipPath>
+        </defs>
+      </svg>
+
+
+      <div className="absolute h-screen w-full">
         <div className="text-center flex justify-center items-center w-full h-full">
           <div className="overlayed">
             <h2 className="text-white text-4xl font-bold drop-shadow-lg">
