+++ src/components/hero/index.tsx
@@ -1,49 +1,171 @@
-import React from "react";
-import Sun from "../../assets/hero/sun.svg";
-import Sea from "../../assets/hero/sea.svg";
-import { ReactComponent as MidRockLeft } from "../../assets/hero/midRockLeft.svg";
-import { ReactComponent as MidRockRight } from "../../assets/hero/midRockRight.svg";
-import { ReactComponent as BaseRock } from "../../assets/hero/baseRock.svg";
-import "../common.scss"
+import React, { useEffect } from "react";
+import Rellax from "rellax";
+import "../common.scss";
 
 export const HeroPage: React.FC = () => {
+  useEffect(() => {
+    new Rellax(".sun", {
+      speed: -1,
+      center: false,
+      round: true,
+      vertical: true,
+      horizontal: false,
+    });
+      new Rellax(".mtn_front", {
+          speed: 3,
+          center: false,
+          round: true,
+          vertical: true,
+          horizontal: false,
+      });
+      new Rellax(".mtn_rear", {
+          speed: -1,
+          center: false,
+          round: true,
+          vertical: true,
+          horizontal: false,
+      });
+      new Rellax(".ocean", {
+          speed: -1,
+          center: false,
+          round: true,
+          vertical: true,
+          horizontal: false,
+      });
+      new Rellax(".information", {
+          speed: -8,
+      });
+  },[]);
   return (
-    <div className="bg-background1 h-screen">
-      <div className="sea-sunset">
-        <div className="sun justify-center w-full absolute h-screen w-full">
-            <img src={Sun} className="mx-auto h-full w-full object-cover"/>
+    <div className="bg-background1 h-screen m-0 grid align-center overflow-hidden relative">
+      <svg
+        viewBox="0 0 1920 1080"
+        fill="none"
+        xmlns="http://www.w3.org/2000/svg"
+        className="absolute"
+      >
+        <g clipPath="url(#clip0)">
+          <g className="sun">
+            <path
+              opacity="0.11"
+              d="M975.5 1732C1567.83 1732 2048 1252.05 2048 660C2048 67.951 1567.83 -412 975.5 -412C383.175 -412 -97 67.951 -97 660C-97 1252.05 383.175 1732 975.5 1732Z"
+              fill="#F28F46"
+            />
+            <path
+              opacity="0.21"
+              d="M975.5 1624.21C1508.27 1624.21 1940.17 1192.52 1940.17 660C1940.17 127.484 1508.27 -304.206 975.5 -304.206C442.724 -304.206 10.8242 127.484 10.8242 660C10.8242 1192.52 442.724 1624.21 975.5 1624.21Z"
+              fill="#F28F46"
+            />
+            <path
+              opacity="0.21"
+              d="M975.5 1509.36C1444.83 1509.36 1825.3 1129.09 1825.3 660C1825.3 190.914 1444.83 -189.355 975.5 -189.355C506.166 -189.355 125.695 190.914 125.695 660C125.695 1129.09 506.166 1509.36 975.5 1509.36Z"
+              fill="#F28F46"
+            />
+            <path
+              opacity="0.21"
+              d="M975.5 1427.28C1399.47 1427.28 1743.16 1083.76 1743.16 660C1743.16 236.243 1399.47 -107.279 975.5 -107.279C551.533 -107.279 207.84 236.243 207.84 660C207.84 1083.76 551.533 1427.28 975.5 1427.28Z"
+              fill="#F28F46"
+            />
+            <path
+              opacity="0.21"
+              d="M975.5 1315.49C1337.72 1315.49 1631.35 1022.02 1631.35 660C1631.35 297.982 1337.72 4.50699 975.5 4.50699C613.284 4.50699 319.65 297.982 319.65 660C319.65 1022.02 613.284 1315.49 975.5 1315.49Z"
+              fill="#F28F46"
+            />
+            <path
+              opacity="0.21"
+              d="M975.5 1224.69C1287.54 1224.69 1540.49 971.87 1540.49 660C1540.49 348.131 1287.54 95.311 975.5 95.311C663.465 95.311 410.51 348.131 410.51 660C410.51 971.87 663.465 1224.69 975.5 1224.69Z"
+              fill="#F28F46"
+            />
+            <path
+              opacity="0.21"
+              d="M975.5 1142.61C1242.17 1142.61 1458.35 926.54 1458.35 660C1458.35 393.46 1242.17 177.386 975.5 177.386C708.831 177.386 492.652 393.46 492.652 660C492.652 926.54 708.831 1142.61 975.5 1142.61Z"
+              fill="#F28F46"
+            />
+            <path
+              opacity="0.21"
+              d="M975.5 1086.35C1211.09 1086.35 1402.07 895.47 1402.07 660C1402.07 424.534 1211.09 233.651 975.5 233.651C739.912 233.651 548.93 424.534 548.93 660C548.93 895.47 739.912 1086.35 975.5 1086.35Z"
+              fill="#F28F46"
+            />
+            <path
+              opacity="0.21"
+              d="M975.5 994.34C1160.24 994.34 1310.01 844.65 1310.01 660C1310.01 475.35 1160.24 325.662 975.5 325.662C790.756 325.662 640.992 475.35 640.992 660C640.992 844.65 790.756 994.34 975.5 994.34Z"
+              fill="#F28F46"
+            />
+            <path
+              d="M975.5 913.28C1115.44 913.28 1228.88 799.88 1228.88 660C1228.88 520.115 1115.44 406.716 975.5 406.716C835.559 406.716 722.115 520.115 722.115 660C722.115 799.88 835.559 913.28 975.5 913.28Z"
+              fill="#FAFF3B"
+            />
+          </g>
+            <g opacity="0.8" className="ocean">
+                <path opacity="0.8" d="M1215.69 849.473H1.09766V849.612C98.9006 856.364 292.396 869.871 522.883 885.995C1005.48 919.623 1650.24 964.69 1918.9 983.569V1002.87C1798.04 1012.93 1600.99 1029.33 1376.84 1048.07C864.941 1090.93 211.653 1145.92 1.09766 1163.97V1665.49H1920V975.164L1215.69 849.473Z" fill="#C94331"/>
+                <path opacity="0.8" d="M1920 643.436H1.09766V1665.49H1920V643.436Z" fill="#ED583B"/>
+                <path opacity="0.8" d="M1719.66 768.437L1128.45 725.988C1128.53 725.851 1128.62 725.713 1128.62 725.574C1129.38 724.335 1130.06 723.095 1130.82 721.717C1131.24 720.889 1131.66 720.201 1132.17 719.373C1133.1 717.72 1133.94 716.204 1134.79 714.551C1135.21 713.723 1135.63 712.896 1136.05 712.207C1136.82 710.829 1137.58 709.313 1138.25 707.934C1138.67 707.108 1139.09 706.281 1139.51 705.453C1140.19 704.076 1140.87 702.698 1141.54 701.319C1141.88 700.631 1142.22 699.94 1142.56 699.252C1143.48 697.322 1144.42 695.392 1145.26 693.463C1145.68 692.637 1146.02 691.672 1146.44 690.845C1147.03 689.468 1147.63 688.226 1148.22 686.848C1148.56 686.022 1148.98 685.195 1149.31 684.23C1149.99 682.576 1150.66 681.06 1151.34 679.407L1536.3 643.573H1.09766V726.953C167.481 725.851 619.927 715.653 809.706 711.242L831.242 709.175C831.58 709.726 781.71 711.891 832.34 710.691C720.134 778.911 895.25 815.782 943.731 815.782C988.154 815.782 1052.27 853.056 1085.21 812.676L1128.62 725.988C1128.28 726.539 1127.94 727.092 1127.61 727.643C1126.6 729.434 1125.49 731.226 1124.48 732.88C1124.06 733.568 1123.64 734.258 1123.22 734.947C1122.46 736.188 1121.69 737.427 1120.93 738.669C1120.51 739.357 1120.01 740.046 1119.59 740.736C1118.49 742.389 1117.39 744.181 1116.29 745.835C1115.95 746.386 1115.53 746.937 1115.2 747.488C1114.34 748.729 1113.5 749.969 1112.58 751.21C1112.07 751.899 1111.65 752.587 1111.13 753.138C1110.21 754.517 1109.2 755.758 1108.27 757.135C1107.93 757.686 1107.5 758.1 1107.17 758.651C1106.07 760.167 1104.98 761.545 1103.79 763.061C1103.28 763.612 1102.86 764.303 1102.35 764.854C1101.43 765.956 1100.5 767.195 1099.65 768.299C1099.23 768.851 1098.81 769.264 1098.47 769.816C1097.21 771.331 1095.93 772.847 1094.75 774.224C1094.33 774.777 1093.91 775.189 1093.4 775.74C1092.47 776.844 1091.54 777.946 1090.53 778.911C1090.11 779.462 1089.6 779.876 1089.18 780.427C1087.91 781.806 1086.56 783.322 1085.21 784.699C1084.87 785.112 1084.53 785.389 1084.11 785.801C1083.1 786.905 1082 788.007 1080.9 789.109C1080.4 789.523 1079.98 790.074 1079.47 790.488C1078.28 791.727 1077.1 792.829 1075.84 793.933C1075.59 794.208 1075.33 794.485 1075.07 794.622C1073.72 795.863 1072.38 797.103 1071.02 798.344C1070.51 798.756 1070.09 799.17 1069.59 799.584C1068.49 800.548 1067.39 801.513 1066.21 802.478C1065.87 802.753 1065.45 803.167 1065.11 803.443C1063.67 804.682 1062.24 805.785 1060.8 806.889C1060.38 807.163 1059.95 807.577 1059.53 807.852C1058.44 808.679 1057.25 809.644 1056.15 810.472C1055.73 810.746 1055.23 811.16 1054.81 811.437C1054.22 811.849 1053.62 812.262 1053.03 812.676L873.64 817.775L257.514 835.278L1.09766 786.905V849.473H1215.7L1719.66 768.437Z" fill="#C94331"/>
+                <path opacity="0.8" d="M1918.9 782.769L1719.66 768.437L1215.7 849.473L1920 975.164V736.325L1918.9 736.462V782.769Z" fill="#EA6E5E"/>
+                <path opacity="0.8" d="M1918.9 1151.02V1002.87C1798.04 1012.93 1604.37 1029.33 1380.31 1048.07L1918.9 1151.02Z" fill="#E37A76"/>
+                <path opacity="0.8" d="M1403.36 1046.14C1395.68 1046.83 1387.99 1047.38 1380.31 1048.07L1918.9 1151.02V1071.64L1403.36 1046.14Z" fill="#AA6465"/>
+                <path opacity="0.8" d="M1376.84 1048.07C1600.99 1029.33 1798.04 1012.93 1918.9 1002.87V983.569C1650.24 964.69 1005.48 919.623 522.883 885.995L1376.84 1048.07Z" fill="#FF833E"/>
+                <path opacity="0.8" d="M1725.74 1018.99L705.568 920.725L1376.84 1048.21C1505.06 1037.32 1624.39 1027.4 1725.74 1018.99Z" fill="#E87338"/>
+                <path opacity="0.8" d="M1163.34 648.121C1162.83 649.637 1162.32 651.153 1161.73 652.669C1161.39 653.634 1161.05 654.599 1160.71 655.564C1160.12 657.08 1159.62 658.596 1159.03 660.249C1158.69 661.214 1158.27 662.316 1157.93 663.281C1157.25 665.073 1156.58 666.864 1155.9 668.656C1155.57 669.621 1155.15 670.586 1154.81 671.412C1154.21 672.79 1153.7 674.169 1153.11 675.548C1152.78 676.373 1152.44 677.201 1152.1 678.166C1151.94 678.579 1151.77 678.993 1151.6 679.268L1556.24 641.506H1165.45C1165.03 642.746 1164.6 644.124 1164.18 645.366C1163.92 646.331 1163.67 647.156 1163.34 648.121Z" fill="#EF763A"/>
+                <path opacity="0.8" d="M522.884 885.995C291.551 869.871 97.4655 856.364 0 849.473L1.09797 1163.97C210.555 1145.92 864.097 1090.93 1376.84 1048.07L522.884 885.995Z" fill="#E37A76"/>
+                <path opacity="0.8" d="M0 968.961L1.09797 1163.97C210.555 1145.92 864.097 1090.93 1376.84 1048.07L1329.46 1039.11C977.524 1020.37 185.134 978.333 0 968.961Z" fill="#AA6465"/>
+                <path opacity="0.8" d="M434.901 707.041C433.594 708.26 432.288 709.377 430.981 710.596C430.328 711.204 429.511 711.814 428.859 712.524C427.715 713.54 426.572 714.453 425.428 715.468C424.612 716.076 423.958 716.787 423.142 717.398C421.346 718.818 419.712 720.239 417.916 721.659C417.263 722.169 416.609 722.676 415.957 723.183C414.65 724.197 413.343 725.214 412.037 726.228C411.22 726.838 410.404 727.447 409.587 728.055C408.117 729.071 406.81 730.187 405.341 731.203C404.524 731.812 403.708 732.422 402.891 732.929C401.258 734.147 399.461 735.364 397.828 736.481C397.011 737.09 396.195 737.599 395.215 738.207C393.745 739.121 392.439 740.137 390.969 741.05C390.806 741.151 390.643 741.252 390.643 741.354L1919.13 783.179V679.125H1217.85L434.901 707.041Z" fill="#F28F46"/>
+                <path opacity="0.8" d="M1782.88 698.735C1780.74 696.828 1778.61 694.921 1776.66 693.015C1775.7 692.196 1774.92 691.243 1773.96 690.425C1772.59 689.063 1771.24 687.837 1769.88 686.476C1769.11 685.658 1768.13 684.84 1767.36 684.024C1765.41 682.115 1763.48 680.073 1761.54 678.166C1760.76 677.348 1759.99 676.53 1759.21 675.576C1757.86 674.215 1756.69 672.853 1755.33 671.49C1754.56 670.536 1753.59 669.583 1752.81 668.765C1751.26 666.995 1749.71 665.224 1748.16 663.452C1747.38 662.499 1746.41 661.41 1745.63 660.455C1744.28 658.958 1743.11 657.459 1741.76 655.825C1740.98 654.871 1740.2 653.916 1739.43 652.963C1738.26 651.464 1737.1 649.967 1735.74 648.468C1735.16 647.65 1734.38 646.697 1733.8 645.879C1732.83 644.654 1731.86 643.291 1730.9 642.065H0.224609V645.335L1784.62 700.506C1784.04 699.826 1783.46 699.28 1782.88 698.735Z" fill="#EF763A"/>
+                <path opacity="0.8" d="M1628.93 816.801C1628.93 816.801 1628.77 816.801 1628.93 816.801C1626.41 816.297 1624.04 815.796 1621.84 815.208C1621.21 815.041 1620.58 814.957 1619.95 814.789C1617.11 814.118 1614.44 813.53 1611.6 812.861C1610.81 812.693 1609.87 812.442 1609.08 812.274C1607.04 811.77 1604.83 811.267 1602.78 810.68C1601.99 810.512 1601.2 810.262 1600.42 810.094C1597.74 809.423 1595.06 808.668 1592.38 807.997C1591.75 807.829 1590.96 807.577 1590.34 807.411C1588.29 806.824 1586.24 806.237 1584.04 805.649C1583.09 805.398 1582.31 805.146 1581.36 804.894C1578.84 804.141 1576.32 803.386 1573.8 802.631C1573.32 802.463 1572.86 802.295 1572.38 802.211C1570.18 801.542 1567.81 800.787 1565.61 800.116C1564.67 799.864 1563.88 799.528 1562.94 799.276C1560.88 798.607 1558.99 797.936 1556.95 797.265C1556.32 797.013 1555.68 796.845 1555.06 796.594C1552.54 795.756 1550.17 794.917 1547.65 793.995C1546.86 793.659 1545.92 793.407 1545.13 793.073C1543.4 792.402 1541.51 791.731 1539.78 791.144C1538.99 790.808 1538.2 790.556 1537.25 790.223C1534.9 789.3 1532.53 788.461 1530.33 787.538C1529.54 787.204 1528.75 786.952 1528.12 786.617C1526.39 785.946 1524.66 785.191 1522.92 784.519C1521.98 784.186 1521.19 783.766 1520.25 783.43C1518.2 782.508 1515.99 781.669 1513.95 780.748C1513.31 780.412 1512.53 780.16 1511.9 779.825C1510.17 778.986 1508.28 778.232 1506.55 777.393C1505.6 776.974 1504.82 776.554 1503.87 776.219C1502.29 775.465 1500.56 774.71 1498.98 773.955C1498.35 773.62 1497.56 773.284 1496.94 772.95C1494.89 771.944 1492.84 770.937 1490.8 769.848C1490.01 769.428 1489.06 769.009 1488.28 768.589C1486.85 767.834 1485.44 767.081 1484.02 766.326C1483.23 765.906 1482.44 765.487 1481.66 765.067C1479.61 763.978 1477.72 762.972 1475.83 761.881C1475.05 761.461 1474.41 761.044 1473.62 760.624C1472.21 759.869 1470.95 759.114 1469.53 758.277C1468.75 757.774 1467.96 757.354 1467.17 756.851C1465.44 755.844 1463.86 754.839 1462.13 753.832C1461.81 753.664 1461.49 753.497 1461.18 753.245L0 797.684L479.92 827.367L1628.93 816.801Z" fill="#FF833E"/>
+                <path opacity="0.8" d="M0 786.766V842.858L257.345 835.555L0 786.766Z" fill="#F28F46"/>
+                <path opacity="0.8" d="M1103.96 683.816L859.113 706.694L1077.78 722.405C1087.15 710.828 1095.93 697.873 1103.96 683.816Z" fill="#F28F46"/>
+            </g>
+            <g className="mtn_rear">
+
+                <path
+                    d="M57.5129 751.594C70.9276 755.601 80.2562 752.596 94.1334 750.67C124.433 746.432 154.731 742.195 185.03 737.957C200.449 748.436 223.655 749.206 242.313 748.513C249.174 748.282 256.036 747.896 262.82 747.434C288.107 745.816 313.241 742.966 338.22 738.804C351.635 744.044 372.219 738.573 384.94 731.716C397.661 724.859 409.996 716.307 424.336 714.766C388.101 699.665 350.015 677.321 310.851 680.172C294.352 681.405 278.085 685.796 261.586 686.182C241.773 686.721 222.345 681.559 202.839 678.092C155.579 669.693 106.623 670.927 59.9029 681.79C40.0891 686.413 20.1984 692.731 0.230469 696.891V753.521C19.5045 757.064 38.9327 757.604 57.5129 751.594Z"
+                    fill="#C94331"
+                />
+                <path
+                    d="M98.8363 734.105C150.259 726.4 204.613 711.915 256.575 713.379C274.307 713.842 291.5 719.158 309.155 721.161C326.81 723.164 346.007 721.392 359.344 709.758C379.929 712.223 401.747 713.071 422.486 713.765C411.923 691.652 377.076 686.104 358.419 670.31C331.975 647.966 301.831 614.451 298.13 580.011C274.615 565.217 248.017 533.551 224.966 518.065C217.642 513.134 201.914 502.579 193.048 502.81C169.226 503.349 145.248 495.337 126.514 480.543C120.038 475.458 109.552 456.813 112.097 448.954C94.3652 437.396 87.349 407.503 78.483 388.241C71.3902 372.986 60.5968 351.413 53.5809 336.08C41.1685 334.231 27.0599 323.598 26.1348 311.117C25.518 302.256 19.5816 287.849 12.1804 282.994C8.40266 280.606 4.39366 278.603 0.230469 277.063V739.113C32.765 742.118 66.3018 738.959 98.8363 734.105Z"
+                    fill="#5F031B"
+                />
+                <path
+                    d="M278.24 694.04C258.504 689.726 236.223 674.624 218.953 663.992C209.702 658.29 197.366 661.064 187.498 656.518C184.722 655.209 182.255 653.436 179.557 652.05C174.777 649.584 169.458 648.582 164.215 647.427C150.414 644.345 136.845 640.03 123.739 634.637C118.882 632.634 113.948 630.322 110.17 626.624C106.393 622.926 101.612 602.662 102.847 597.5C91.8986 585.558 82.8783 571.921 76.1709 557.127C71.0055 545.648 69.8491 521.455 66.0714 509.435C58.0534 502.347 60.8288 489.172 56.2802 479.465C52.5795 471.529 44.4845 466.751 37.9313 460.974C31.3781 455.118 26.5211 442.867 30.2988 435.008C26.2898 405.345 22.2808 375.606 18.3488 345.942C15.6506 325.756 12.875 305.261 5.70511 286.153C4.62577 283.38 3.46933 280.529 2.23578 277.832C1.46483 277.524 0.770961 277.217 0 276.908V738.958C32.6116 741.963 66.1484 738.804 98.6826 733.95C150.106 726.246 204.459 711.761 256.422 713.225C274.154 713.688 291.346 719.004 309.001 721.007C319.1 722.162 329.818 722.085 339.531 719.542C320.026 708.833 299.981 698.817 278.24 694.04Z"
+                    fill="#470318"
+                />
+                <path
+                    d="M1871.9 750.578C1856.87 755.072 1846.42 751.701 1830.87 749.54C1796.93 744.785 1763 740.031 1729.06 735.277C1711.78 747.033 1685.79 747.898 1664.89 747.12C1657.21 746.86 1649.52 746.428 1641.92 745.91C1613.59 744.094 1585.44 740.895 1557.46 736.227C1542.43 742.105 1519.37 735.969 1505.12 728.275C1490.96 720.582 1477.05 710.986 1460.99 709.258C1501.58 692.314 1544.24 667.246 1588.11 670.444C1606.6 671.827 1624.82 676.755 1643.3 677.187C1665.49 677.792 1687.25 672 1709.1 668.111C1762.05 658.688 1816.88 660.071 1869.22 672.259C1891.42 677.447 1897.46 677.014 1919.83 681.768V745.305C1898.15 749.194 1892.71 757.319 1871.9 750.578Z"
+                    fill="#C94331"
+                />
+                <path
+                    d="M1809.28 730.954C1751.68 722.31 1690.8 706.059 1632.59 707.701C1612.73 708.22 1593.47 714.184 1573.69 716.432C1553.91 718.68 1532.41 716.692 1517.47 703.638C1494.41 706.405 1469.97 707.356 1446.74 708.134C1458.57 683.325 1497.61 677.1 1518.51 659.38C1548.13 634.311 1581.89 596.708 1586.04 558.068C1612.38 541.471 1642.18 505.943 1668 488.568C1676.2 483.035 1693.82 471.193 1703.75 471.452C1730.43 472.057 1757.3 463.067 1778.28 446.47C1785.54 440.765 1797.28 419.845 1794.43 411.028C1814.3 398.061 1822.16 364.522 1832.08 342.911C1840.03 325.709 1852.03 301.504 1859.98 284.302C1873.88 282.227 1889.69 270.298 1890.72 256.294C1891.41 246.353 1898.06 230.189 1906.36 224.743C1910.58 222.063 1915.08 219.815 1919.74 218.087V736.487C1883.3 739.945 1845.73 736.401 1809.28 730.954Z"
+                    fill="#5F031B"
+                />
+                <path
+                    d="M1608.32 686.004C1630.43 681.163 1655.39 664.221 1674.73 652.291C1685.1 645.894 1698.92 649.006 1709.97 643.906C1713.08 642.436 1715.84 640.448 1718.86 638.893C1724.22 636.126 1730.17 635.003 1736.05 633.705C1751.5 630.248 1766.7 625.407 1781.39 619.356C1786.83 617.108 1792.36 614.515 1796.59 610.366C1800.82 606.217 1806.18 583.482 1804.79 577.691C1817.05 564.292 1827.16 548.991 1834.68 532.394C1840.46 519.514 1841.76 492.371 1845.99 478.885C1854.97 470.933 1851.86 456.152 1856.96 445.259C1861.1 436.355 1870.17 430.996 1877.51 424.513C1884.85 417.943 1890.29 404.199 1886.06 395.382C1890.55 362.101 1895.04 328.734 1899.45 295.453C1902.46 272.805 1905.58 249.811 1913.61 228.373C1914.82 225.261 1916.11 222.063 1917.49 219.037C1918.36 218.692 1919.14 218.346 1920 218V736.4C1883.47 739.772 1845.9 736.227 1809.45 730.781C1751.85 722.137 1690.97 705.886 1632.76 707.528C1612.9 708.047 1593.64 714.012 1573.86 716.259C1562.55 717.555 1550.54 717.469 1539.66 714.617C1561.51 702.601 1583.97 691.364 1608.32 686.004Z"
+                    fill="#470318"
+                />
+            </g>
+          <g className="mtn_front">
+              <path d="M1955.48 149.621L1924.53 148.33C1924.53 148.33 1865.15 141.881 1861.29 143.171C1857.42 144.461 1813.51 145.75 1810.02 148.33C1806.53 150.911 1792.73 152.657 1785.15 154.781C1777.56 156.906 1765.81 159.942 1765.81 159.942H1754.2C1754.2 159.942 1740.02 166.392 1736.76 167.682C1733.58 168.972 1707.71 179.294 1707.71 179.294L1672.9 180.584L1651.44 178.004L1601.92 175.423L1583.87 166.392L1565.82 159.942L1547.77 165.102L1523.27 175.423H1503.94H1487.41H1463.97H1450.09L1442.66 187.034L1450.09 201.301L1463.97 210.332L1478.23 219.364L1482.48 224.22V227.863V240.308C1482.48 240.308 1483.99 247.974 1484.44 250.326C1484.9 252.679 1486.8 268.919 1486.8 268.919C1486.8 268.919 1486.8 285.159 1486.8 286.601C1486.8 288.043 1487.25 310.506 1486.8 312.404C1486.34 314.301 1483.54 330.541 1484.67 332.438C1485.81 334.335 1486.95 328.72 1486.95 328.72L1489 320.676C1489 320.676 1490.52 315.059 1490.52 314.149C1490.52 313.238 1490.82 308.002 1490.52 307.319C1490.21 306.636 1491.05 302.538 1490.67 301.324C1490.21 300.11 1490.82 293.887 1490.82 293.887C1490.82 293.887 1490.75 288.27 1490.82 287.36C1490.9 286.449 1490.82 281.137 1490.82 281.137C1490.82 281.137 1490.52 272.486 1490.82 271.651C1491.12 270.74 1490.21 268.084 1490.82 267.173C1491.42 266.263 1490.67 262.696 1490.9 261.785C1491.12 260.874 1491.57 255.942 1491.05 255.259C1490.52 254.651 1491.27 250.174 1491.27 250.174L1494.08 253.437V258.218V262.392V267.477V271.044L1497.04 270.133C1497.04 270.133 1497.04 263.607 1497.04 262.392C1497.04 261.178 1496.74 259.13 1497.04 257.611C1497.34 256.093 1498.55 254.651 1498.85 253.437C1499.16 252.223 1498.55 251.388 1498.85 249.87C1499.16 248.353 1498.7 246.607 1498.93 245.393C1499.16 244.179 1498.93 242.433 1498.93 242.433C1498.93 242.433 1497.19 242.13 1497.57 239.929L1505.46 247.063L1507.5 248.504L1506.44 253.741L1506.97 257.839C1506.97 257.839 1507.42 261.709 1507.58 263.227C1507.73 264.745 1508.26 267.932 1508.33 269.45C1508.41 270.968 1508.33 277.798 1508.33 277.798V283.186V289.712V298.06V303.145C1508.33 303.145 1505.38 304.967 1504.78 307.319C1504.17 309.672 1502.72 311.72 1502.42 313.845C1502.12 315.895 1502.12 317.109 1502.12 318.323C1502.12 319.537 1502.42 320.978 1502.42 321.89C1502.42 322.8 1503.03 324.546 1503.03 324.546C1502.42 321.89 1505.08 318.626 1505.38 317.716C1505.69 316.805 1505.69 314.756 1506.28 313.542C1506.59 313.011 1507.42 311.492 1508.26 310.202C1509.17 308.761 1510.01 307.395 1510.01 307.395C1510.01 307.395 1512.21 304.055 1512.21 303.221C1512.21 302.386 1512.21 299.958 1512.21 298.136C1512.21 296.315 1511.9 294.569 1512.21 293.052C1512.51 291.534 1511.6 288.878 1512.21 287.132C1512.81 285.311 1511.14 283.262 1512.28 282.048C1513.42 280.834 1512.81 279.088 1513.72 278.177C1514.63 277.267 1515.54 274.307 1515.54 274.307C1515.54 274.307 1516.15 272.258 1515.54 271.044C1514.93 269.83 1514.63 268.691 1514.32 267.477C1514.02 266.263 1513.12 264.821 1513.12 263.607C1513.12 262.392 1512.51 261.558 1512.51 259.736C1512.51 257.915 1512.28 251.996 1512.28 251.996L1516.91 255.335L1532.45 258.067C1532.45 262.544 1532.45 267.629 1532.45 268.388C1532.45 269.83 1532.91 292.293 1532.45 294.19C1532 296.087 1529.19 312.328 1530.33 314.225C1531.47 316.122 1532.61 310.506 1532.61 310.506L1534.66 302.462C1534.66 302.462 1536.16 296.846 1536.16 295.935C1536.16 295.025 1536.47 289.788 1536.16 289.106C1535.86 288.422 1536.7 284.325 1536.32 283.11C1535.86 281.896 1536.47 275.673 1536.47 275.673C1536.47 275.673 1536.39 270.058 1536.47 269.146C1536.54 268.236 1536.47 262.923 1536.47 262.923C1536.47 262.923 1536.39 260.874 1536.39 258.75L1554.07 261.861V265.049V271.575V279.922V285.007C1554.07 285.007 1551.1 286.829 1550.51 289.182C1549.9 291.534 1548.46 293.583 1548.15 295.707C1547.85 297.757 1547.85 298.971 1547.85 300.186C1547.85 301.4 1548.15 302.841 1548.15 303.752C1548.15 304.663 1548.76 306.408 1548.76 306.408C1548.15 303.752 1550.8 300.488 1551.1 299.578C1551.41 298.668 1551.41 296.619 1552.02 295.405C1552.32 294.873 1553.16 293.355 1553.99 292.065C1554.9 290.624 1555.74 289.258 1555.74 289.258C1555.74 289.258 1557.94 285.918 1557.94 285.083C1557.94 284.249 1557.94 281.82 1557.94 279.998C1557.94 278.177 1557.64 276.432 1557.94 274.915C1558.23 273.396 1557.33 270.74 1557.94 268.994C1558.54 267.249 1556.88 265.125 1558.01 263.91C1558.31 263.531 1558.54 263.151 1558.61 262.696L1616.71 272.865H1679.2H1743.06L1810.25 307.395V314.68C1810.25 314.68 1811.76 322.345 1812.22 324.697C1812.67 327.05 1814.57 343.29 1814.57 343.29C1814.57 343.29 1814.57 359.531 1814.57 360.972C1814.57 362.414 1815.03 384.878 1814.57 386.775C1814.11 388.672 1811.3 404.912 1812.44 406.809C1813.58 408.707 1814.72 403.091 1814.72 403.091L1816.77 395.047C1816.77 395.047 1818.28 389.431 1818.28 388.521C1818.28 387.609 1818.59 382.374 1818.28 381.69C1817.99 381.084 1818.81 376.909 1818.43 375.695C1818.06 374.481 1818.59 368.258 1818.59 368.258C1818.59 368.258 1818.51 362.642 1818.59 361.732C1818.66 360.821 1818.59 355.509 1818.59 355.509C1818.59 355.509 1818.28 346.857 1818.59 346.023C1818.89 345.111 1817.99 342.456 1818.59 341.544C1819.19 340.634 1818.43 337.067 1818.66 336.157C1818.89 335.246 1819.34 330.314 1818.81 329.63C1818.28 328.947 1819.04 324.546 1819.04 324.546L1821.85 327.809V332.59V336.764V341.848V345.415L1824.81 344.505C1824.81 344.505 1824.81 337.978 1824.81 336.764C1824.81 335.549 1824.51 333.501 1824.81 331.983C1825.12 330.465 1826.32 329.023 1826.62 327.809C1826.93 326.595 1826.32 325.76 1826.62 324.242C1826.93 322.724 1826.47 320.978 1826.7 319.764C1826.85 319.006 1826.78 317.943 1826.78 317.336L1835.35 322.496L1834.21 328.113L1834.74 332.21C1834.74 332.21 1835.2 336.081 1835.35 337.599C1835.5 339.116 1836.03 342.304 1836.11 343.822C1836.18 345.339 1836.11 352.17 1836.11 352.17V357.557V364.084V372.432V377.517C1836.11 377.517 1833.15 379.338 1832.55 381.69C1831.94 384.043 1830.5 386.092 1830.19 388.217C1829.89 390.266 1829.89 391.48 1829.89 392.694C1829.89 393.908 1830.19 395.35 1830.19 396.261C1830.19 397.171 1830.8 398.917 1830.8 398.917C1830.19 396.261 1832.85 392.998 1833.15 392.088C1833.45 391.176 1833.45 389.127 1834.06 387.913C1834.36 387.382 1835.2 385.865 1836.03 384.574C1836.94 383.132 1837.78 381.766 1837.78 381.766C1837.78 381.766 1839.98 378.427 1839.98 377.593C1839.98 376.757 1839.98 374.329 1839.98 372.508C1839.98 370.686 1839.68 368.941 1839.98 367.423C1840.28 365.905 1839.37 363.249 1839.98 361.504C1840.58 359.758 1838.92 357.633 1840.06 356.419C1841.19 355.205 1840.58 353.46 1841.49 352.549C1842.4 351.638 1843.31 348.679 1843.31 348.679C1843.31 348.679 1843.92 346.629 1843.31 345.415C1842.7 344.201 1842.4 343.063 1842.1 341.848C1841.79 340.634 1840.88 339.192 1840.88 337.978C1840.88 336.764 1840.28 335.929 1840.28 334.108C1840.28 332.286 1840.06 326.367 1840.06 326.367C1840.06 326.367 1840.06 325.987 1839.98 325.381L1843.46 327.505V336.991V349.437C1843.46 349.437 1844.98 357.103 1845.44 359.455C1845.89 361.808 1847.79 378.047 1847.79 378.047C1847.79 378.047 1847.79 394.288 1847.79 395.73C1847.79 397.171 1848.24 419.635 1847.79 421.532C1847.34 423.43 1844.53 439.67 1845.67 441.567C1846.8 443.464 1847.93 437.849 1847.93 437.849L1849.98 429.804C1849.98 429.804 1851.5 424.188 1851.5 423.278C1851.5 422.367 1851.81 417.131 1851.5 416.447C1851.2 415.841 1852.03 411.666 1851.65 410.452C1851.27 409.238 1851.81 403.015 1851.81 403.015C1851.81 403.015 1851.73 397.399 1851.81 396.489C1851.88 395.578 1851.81 390.266 1851.81 390.266C1851.81 390.266 1851.5 381.614 1851.81 380.78C1852.11 379.869 1851.2 377.213 1851.81 376.303C1852.41 375.391 1851.65 371.824 1851.88 370.914C1852.11 370.004 1852.57 365.071 1852.03 364.388C1851.5 363.704 1852.26 359.303 1852.26 359.303L1855.07 362.566V367.347V371.522V376.605V380.172L1858.02 379.262C1858.02 379.262 1858.02 372.736 1858.02 371.522C1858.02 370.307 1857.72 368.258 1858.02 366.741C1858.33 365.223 1859.54 363.78 1859.85 362.566C1860.15 361.352 1859.54 360.517 1859.85 358.999C1860.15 357.481 1859.69 355.736 1859.92 354.522C1860.15 353.308 1859.92 351.562 1859.92 351.562C1859.92 351.562 1858.02 351.258 1858.63 348.906C1859.24 346.553 1858.63 339.724 1858.63 339.724L1859.31 337.143L1866.52 341.469L1869.33 353.308L1867.43 362.87L1867.96 366.967C1867.96 366.967 1868.42 370.838 1868.57 372.356C1868.72 373.874 1869.25 377.061 1869.33 378.579C1869.4 380.096 1869.33 386.927 1869.33 386.927V392.314V398.841V407.189V412.274C1869.33 412.274 1866.37 414.095 1865.76 416.447C1865.15 418.8 1863.71 420.849 1863.41 422.974C1863.1 425.023 1863.1 426.237 1863.1 427.451C1863.1 428.666 1863.41 430.108 1863.41 431.018C1863.41 431.929 1864.01 433.674 1864.01 433.674C1863.41 431.018 1866.06 427.755 1866.37 426.845C1866.67 425.934 1866.67 423.884 1867.28 422.67C1867.58 422.14 1868.42 420.622 1869.25 419.331C1870.16 417.889 1870.99 416.523 1870.99 416.523C1870.99 416.523 1873.19 413.184 1873.19 412.35C1873.19 411.515 1873.19 409.087 1873.19 407.265C1873.19 405.444 1872.89 403.698 1873.19 402.18C1873.5 400.663 1872.58 398.007 1873.19 396.261C1873.8 394.44 1872.13 392.39 1873.27 391.176C1874.41 389.962 1873.8 388.217 1874.71 387.307C1875.62 386.395 1876.53 383.436 1876.53 383.436C1876.53 383.436 1877.14 381.386 1876.53 380.172C1875.93 378.958 1875.62 377.82 1875.32 376.605C1875.01 375.391 1874.1 373.95 1874.1 372.736C1874.1 371.522 1873.5 370.686 1873.5 368.865C1873.5 367.043 1873.27 361.124 1873.27 361.124C1873.27 361.124 1872.89 355.129 1873.5 353.08C1873.8 351.942 1873.88 348.603 1873.88 345.947L1876.76 347.692V356.647C1876.76 356.647 1878.27 364.312 1878.73 366.665C1879.18 369.017 1881.08 385.257 1881.08 385.257C1881.08 385.257 1881.08 401.498 1881.08 402.94C1881.08 404.381 1881.53 426.845 1881.08 428.741C1880.62 430.639 1877.81 446.879 1878.95 448.777C1880.09 450.674 1881.23 445.058 1881.23 445.058L1883.28 437.013C1883.28 437.013 1884.8 431.398 1884.8 430.487C1884.8 429.577 1885.1 424.34 1884.8 423.657C1884.5 423.05 1885.32 418.876 1884.95 417.661C1884.57 416.447 1885.09 410.225 1885.09 410.225C1885.09 410.225 1885.03 404.609 1885.09 403.698C1885.17 402.788 1885.09 397.475 1885.09 397.475C1885.09 397.475 1884.8 388.824 1885.09 387.989C1885.4 387.079 1884.5 384.422 1885.09 383.512C1885.7 382.601 1884.95 379.034 1885.17 378.123C1885.4 377.213 1885.85 372.28 1885.32 371.597C1884.8 370.914 1885.55 366.513 1885.55 366.513L1888.36 369.776V374.557V378.731V383.815V387.382L1891.32 386.471C1891.32 386.471 1891.32 379.945 1891.32 378.731C1891.32 377.517 1891.02 375.467 1891.32 373.95C1891.62 372.432 1892.83 370.99 1893.13 369.776C1893.44 368.561 1892.83 367.727 1893.13 366.209C1893.44 364.691 1892.98 362.946 1893.21 361.732C1893.44 360.517 1893.21 358.772 1893.21 358.772C1893.21 358.772 1891.7 358.544 1891.78 356.799L1902.08 363.022L1900.64 370.155L1901.17 374.253C1901.17 374.253 1901.63 378.123 1901.78 379.641C1901.93 381.16 1902.46 384.346 1902.54 385.865C1902.62 387.382 1902.54 394.212 1902.54 394.212V399.6V406.127V414.474V419.559C1902.54 419.559 1899.59 421.38 1898.98 423.733C1898.37 426.085 1896.93 428.135 1896.63 430.259C1896.32 432.308 1896.32 433.522 1896.32 434.737C1896.32 435.951 1896.63 437.393 1896.63 438.303C1896.63 439.215 1897.23 440.96 1897.23 440.96C1896.63 438.303 1899.28 435.04 1899.59 434.13C1899.89 433.22 1899.89 431.17 1900.49 429.956C1900.79 429.425 1901.63 427.907 1902.46 426.617C1903.38 425.175 1904.21 423.809 1904.21 423.809C1904.21 423.809 1906.41 420.47 1906.41 419.635C1906.41 418.725 1906.41 416.371 1906.41 414.55C1906.41 412.729 1906.11 410.984 1906.41 409.466C1906.72 407.948 1905.81 405.292 1906.41 403.546C1907.02 401.726 1905.35 399.676 1906.49 398.462C1907.62 397.247 1907.02 395.502 1907.92 394.591C1908.83 393.681 1909.74 390.721 1909.74 390.721C1909.74 390.721 1910.35 388.672 1909.74 387.457C1909.14 386.243 1908.83 385.105 1908.53 383.891C1908.23 382.677 1907.32 381.235 1907.32 380.021C1907.32 378.807 1906.72 377.971 1906.72 376.151C1906.72 375.543 1906.72 374.481 1906.64 373.342L1939.78 481.788L1923.32 519.809L1890.41 545.156L1835.58 555.552L1769.07 590.613L1753.75 596.154H1747.07L1742.68 608.599L1747.83 624.612L1755.11 631.897L1759.66 644.647V661.115C1759.66 661.115 1754.51 673.864 1753.82 675.23C1753.14 676.596 1742.15 685.551 1742.68 686.614C1743.29 687.676 1732.13 693.443 1732.13 693.443L1724.25 701.488L1721.14 708.773L1719.7 717.197H1707.03L1699.44 723.344L1693.83 732.147L1641.28 741.026L1583.79 819.873L1556.27 892.276L1436.97 964.671L1399.24 1268.69H1978V149.545L1955.48 149.621Z" fill="#33132C"/>
+              <path d="M749.146 950.49C733.525 950.49 700.153 963.029 684.808 960.058C666.849 956.622 641.251 959.441 623.842 953.86C597.35 945.393 554.344 952.208 531.774 935.948C515.947 924.52 497.369 903.648 480.923 893.178C455.533 876.919 430.486 860.175 405.783 842.814C402.961 840.821 390.92 832.763 389.681 829.526C381.562 808.653 362.088 789.224 353.212 768.696C344.266 748.03 335.184 726.813 333.738 704.288C333.188 696.296 330.023 681.21 326.445 673.977C312.339 674.735 304.838 647.181 290.732 648.214C264.034 650.074 228.734 660.476 203.343 652.14C180.085 644.494 156.759 625.55 140.451 607.227C115.61 606.882 84.989 603.507 62.9015 592.003C44.8044 582.565 26.088 567.548 15.4225 550.12C6.06426 534.827 6.27071 517.949 0.146572 501.141C-2.7435 493.219 -5.01424 481.715 -7.8354 473.725C-17.5377 472.691 -27.1711 465.871 -35.5659 460.843C-48.5023 453.058 -56.8283 438.179 -56.6907 423.024C-57.1035 422.887 -57.5164 422.542 -57.998 422.06V1268.69H806.69H1415.69L1454.41 959.947L1261.85 1001.32L1044.35 991.776L815.479 966.265C815.479 966.189 815.409 966.189 815.409 966.123C796.556 960.409 768.894 950.49 749.146 950.49Z" fill="#33132C"/>
+              <path d="M381.766 939.261C379.564 936.433 376.674 934.297 373.784 932.161C362.362 923.694 337.796 909.362 326.373 900.895C325.617 897.582 321.075 897.098 317.703 897.449C303.322 898.826 287.358 895.105 273.046 896.481C263.138 887.255 252.885 863.763 256.807 850.807C259.628 841.372 263 826.498 270.706 820.29C261.486 794.738 249.1 760.57 244.077 733.911C235.406 725.989 216.965 719.513 205.956 723.439C189.373 729.295 172.858 738.457 158.133 747.963C131.71 764.978 82.1661 769.386 63.1748 744.312C55.1238 733.704 43.9083 718.273 37.3019 706.7C27.8063 704.84 15.4893 697.263 5.92464 698.365C-8.59428 683.279 -23.2508 665.161 -31.7145 646.011C-40.4534 648.353 -49.1923 650.626 -58 652.969V1948H781H1212.5H1583H1978C1978 1688.01 1978 862.925 1978 602.93C1978 602.93 1940.87 780.643 1911.69 684.094C1866.7 729.234 1862.89 868.703 1780.13 780.643C1746.7 821.293 1797.64 890.983 1717.53 900.002C1693.99 919.797 1681.46 986.471 1647.51 936.075C1628.41 923.874 1446.99 981.166 1463.43 1024.14C1439.03 1093.63 1213.57 1031.56 1228.43 1084.61C1127.63 1048.54 1103.23 1146.15 1020.48 1114.32C1049.65 1074.53 798.202 1037.4 814.117 1084.61C738.788 1127.58 703.776 1048.54 665.051 1084.61C532.96 1084.61 548.344 1025.2 517.576 1048.54C463.997 1047.48 389.61 956.071 387.477 950.832C385.963 946.769 384.449 942.64 381.766 939.261Z" fill="#280F23"/>
+          </g>
+        </g>
+        <defs>
+          <clipPath id="clip0">
+            <rect width="1920" height="1080" fill="white" />
+          </clipPath>
+        </defs>
+      </svg>
+
+        <div className="absolute top-0 h-screen w-full p-32 information">
+            <div className="space-y-6 overlayed">
+                <h1 className="text-6xl font-bold text-white">Sunset Hills</h1>
+                <div>
+                    <p className="text-2xl font-semibold text-white">when given an array of buildings, and the </p>
+                    <p className="text-2xl font-semibold text-white">sun is to the left side, visualise which </p>
+                    <p className="text-2xl font-semibold text-white">
+                        buildings or it’s parts of buildings are sunlit.
+                    </p>
+                </div>
+            </div>
         </div>
-        <div className="sea flex items-end w-full absolute h-screen">
-            <img src={Sea} className="opacity-90 h-1/2 w-full object-cover"/>
-        </div>
-          <div className="midRock absolute h-screen w-full">
-              <div className="flex h-full items-center">
-                  <div className="my-full h-1/4 inline-block w-full flex space-between">
-                      <MidRockLeft className="h-full inline"/>
-                      <MidRockRight className="h-full inline ml-auto"/>
-                  </div>
-              </div>
-          </div>
-          <div className="baseRock absolute h-screen w-full">
-              <div className="flex h-full items-end">
-                  <div className="my-full h-2/3 w-full flex items-end">
-                      <BaseRock className="w-full inline"/>
-                  </div>
-              </div>
-          </div>
-          <div className="absolute h-screen w-full p-32">
-              <div className="space-y-6 overlayed">
-                  <h1 className="text-6xl font-bold text-white">Sunset Hills</h1>
-                  <div>
-                      <p className="text-2xl font-semibold text-white">when given an array of buildings, and the </p>
-                      <p className="text-2xl font-semibold text-white">sun is to the left side, visualise which </p>
-                      <p className="text-2xl font-semibold text-white">
-                          buildings or it’s parts of buildings are sunlit.
-                      </p>
-                  </div>
-              </div>
-          </div>
-      </div>
     </div>
   );
 };
